+++ shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './constants';
 export * from './cards';
 export * from './validation';
+export * from './deploy';
 export * from './types';
 export * from './protocol';